requirements.txt
snowflake-connector-python==3.0.3
streamlit==1.21.0
pandas==1.5.3
altair==4.2.2
numpy==1.24.3